cuda 7.5 sample ( https://developer.download.nvidia.com/compute/cuda/7.5/Prod/local_installers/cuda_7.5.18_linux.run (driver 352) )

+++ 0_Simple/clock/clock.cu
@@ -18,6 +18,7 @@
 
 // System includes
 #include <stdio.h>
+#include <stdint.h>
 #include <assert.h>
 
 // CUDA runtime
@@ -141,7 +142,7 @@ int main(int argc, char **argv)
         maxEnd = timer[NUM_BLOCKS+i] > maxEnd ? timer[NUM_BLOCKS+i] : maxEnd;
     }
 
-    printf("Total clocks = %d\n", (int)(maxEnd - minStart));
+    printf("Total clocks = %Lf\n", (long double)(maxEnd - minStart));
 
 
     // cudaDeviceReset causes the driver to clean up all state. While
